Reform the project structure into cpp-style project
