--- 2d_space/node_shape.h
@@ -1,22 +1,0 @@
-#pragma once
-
-#include "shape.h"
-
-/**
- * Обыкновенный шейп нода
- */
-struct nodeShape final : public shape2D
-{
-    float radius;
-
-    nodeShape() = delete;
-
-    nodeShape(std::string* label,
-              const point2D initial_point,
-              const int color,
-              const float radius)
-        : shape2D(label, initial_point, color)
-        , radius{radius}
-    {
-    }
-};
