--- graph/edge_data.h
@@ -1,12 +1,0 @@
-#pragma once
-#include <string>
-
-class edgeData
-{
-public:
-    edgeData() = default;
-
-    virtual ~edgeData() noexcept = default;
-
-    virtual const std::string* to_label() const = 0;
-};
