--- graph/edge_type.h
@@ -1,7 +1,0 @@
-#pragma once
-
-enum edgeType
-{
-    DIRECTIONAL,
-    BIDIRECTIONAL,
-};
