+++ include/task1_graph/drawer/drawer_commands/drawer_command_factory.h
@@ -1,9 +1,7 @@
 #pragma once
-#include "drawer_create_command.h"
-#include "drawer_delete_command.h"
-#include "drawer_recolor_command.h"
+#include "drawer_command.h"
+#include "shape.h"
 #include "graph.h"
-#include "node_shape.h"
 #include "algorithm.h"
 #include <cmath>
 
@@ -96,17 +94,10 @@ public:
 
     const drawerCreateCommand* get_node_create_command(const node* node) const
     {
-        // TODO: Описать, как создаются шейпы для наших нодов. Связать логику (нод) и графику (шейп нода)
-        // тут же и опредяляется логика того, какие будут КООРДИНАТЫ и всё такое
-
-        // Движок нужен для того чтобы команда знала где исполняться
         if (engine == nullptr)
             throw 1;
 
         // create 2d objects from 2d_space and supply command with these shapes
-        return new drawerCreateCommand(engine,
-                                       // new nodeShape(...аргументы и факты); и переходишь к nodeShape, используешь соответствующий конструктор. А пока
-                                       nullptr
-        );
+        return nullptr; // new drawerCreateCommand(engine, nullptr);
     }
 };
